+++ runway/hooks/staticsite/auth_at_edge/lambda_config.py
@@ -2,6 +2,8 @@
 import logging
 import os
 import re
+import secrets
+import shutil
 import tempfile
 from distutils.dir_util import copy_tree  # pylint: disable=E
 from tempfile import mkstemp
@@ -44,6 +46,8 @@ def write(
         client_id (str): The ID of the Cognito User Pool Client.
         cookie_settings (dict): The settings for our customized cookies.
         http_headers (dict): The additional headers added to our requests.
+        nonce_signing_secret_param_name (str): SSM param name to store nonce
+            signing secret.
         oauth_scopes (List[str]): The validation scopes for our OAuth requests.
         redirect_path_auth_refresh (str): The URL path for authorization refresh
             redirect (Correlates to the refresh auth lambda).
@@ -51,6 +55,8 @@ def write(
             sign in (Correlates to the parse auth lambda).
         redirect_path_sign_out (str): The URL path to be redirected to after
             sign out (Correlates to the root to be asked to resigning).
+        required_group (Optional[str]): Optional User Pool group to which
+            access should be restricted.
         user_pool_id (str): The ID of the Cognito User Pool.
 
     """
@@ -64,8 +70,13 @@ def write(
         "redirect_path_auth_refresh": kwargs["redirect_path_refresh"],
         "redirect_path_sign_in": kwargs["redirect_path_sign_in"],
         "redirect_path_sign_out": kwargs["redirect_path_sign_out"],
+        "required_group": kwargs.get("required_group"),
         "user_pool_id": context.hook_data["aae_user_pool_id_retriever"]["id"],
     }
+
+    config["nonce_signing_secret"] = get_nonce_signing_secret(
+        kwargs["nonce_signing_secret_param_name"], context,
+    )
 
     # Shared file that contains the method called for configuration data
     path = os.path.join(os.path.dirname(__file__), "templates", "shared.py")
@@ -106,6 +117,12 @@ def write(
                 with open(os.path.join(dirpath, filename), "wb") as newfile:
                     newfile.write(raw.encode())
 
+            # Copy the shared jose-dependent util module to the temporary folder
+            shutil.copyfile(
+                os.path.join(os.path.dirname(__file__), "templates", "shared_jose.py"),
+                os.path.join(dirpath, "shared_jose.py"),
+            )
+
             # Upload our temporary folder to our S3 bucket for
             # Lambda use
             lamb = aws_lambda.upload_lambda_functions(
@@ -126,3 +143,38 @@ def write(
             context_dict.update(lamb)
 
     return context_dict
+
+
+def get_nonce_signing_secret(
+    param_name,  # type: str
+    context,  # type: Context
+):
+    # type: (...) -> str
+    """Retrieve signing secret, generating & storing it first if not present."""
+    session = context.get_session()
+    ssm_client = session.client("ssm")
+    try:
+        response = ssm_client.get_parameter(Name=param_name, WithDecryption=True)
+        return response["Parameter"]["Value"]
+    except ssm_client.exceptions.ParameterNotFound:
+        secret = random_key(16)
+        ssm_client.put_parameter(
+            Description="Auth@Edge nonce signing secret",
+            Name=param_name,
+            Type="String",
+            Value=secret,
+        )
+        return secret
+
+
+def random_key(length=16):
+    """Generate a random key of specified length from the allowed secret characters.
+
+    Args:
+        length (int): The length of the random key.
+
+    """
+    secret_allowed_chars = (
+        "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-._~"
+    )
+    return "".join(secrets.choice(secret_allowed_chars) for _ in range(length))